refactor data types and add score averaging
